requirements.txt
pandas==1.3.3
requests==2.26.0
beautifulsoup4==4.10.0
pygsheets==2.0.5
google-cloud-bigquery==2.26.0
google-play-scraper==1.0.3
app-store-scraper==0.3.5